extract YTDL class from example snippet voice_stream_yt_repeat.py
