+++ examples/voice_stream_yt_repeat.py
@@ -4,53 +4,8 @@ import asyncio
 from typing import Any
 
 import discord
-import youtube_dl
 from discord import Intents
-
-# Suppress noise about console usage from errors
-youtube_dl.utils.bug_reports_message = lambda: ''
-
-ytdl_format_options = {
-    'format': 'bestaudio/best',
-    'outtmpl': '%(extractor)s-%(id)s-%(title)s.%(ext)s',
-    'restrictfilenames': True,
-    'noplaylist': True,
-    'nocheckcertificate': True,
-    'ignoreerrors': False,
-    'logtostderr': False,
-    'quiet': True,
-    'no_warnings': True,
-    'default_search': 'auto',
-    'source_address': '0.0.0.0',  # bind to ipv4 since ipv6 addresses cause issues sometimes
-}
-
-ffmpeg_options = {
-    'options': '-vn',
-}
-
-ytdl = youtube_dl.YoutubeDL(ytdl_format_options)
-
-
-class YTDLSource(discord.PCMVolumeTransformer):
-    def __init__(self, source, *, data, volume=0.5):
-        super().__init__(source, volume)
-
-        self.data = data
-
-        self.title = data.get('title')
-        self.url = data.get('url')
-
-    @classmethod
-    async def from_url(cls, url, *, loop=None, stream=False):
-        loop = loop or asyncio.get_event_loop()
-        data = await loop.run_in_executor(executor=None, func=lambda: ytdl.extract_info(url, download=not stream))
-
-        if 'entries' in data:
-            # take first item from a playlist
-            data = data['entries'][0]
-
-        filename = data['url'] if stream else ytdl.prepare_filename(data)
-        return cls(discord.FFmpegPCMAudio(filename, **ffmpeg_options), data=data)
+from youtube.YTDL import YTDLSource
 
 
 class VoiceJoinerClient(discord.Client):
@@ -150,24 +105,12 @@ def get_url() -> str:
 
     while True:  # I wish was a do-while loop
         url = input("YouTube Video > ")
-        if youtube_url_match(url):
+        if "youtu" in url:
             return url
         print("URL format appears to be incorrect.\nPlease enter a valid YouTube Video URL.")
 
 
-def youtube_url_match(url: str) -> bool:
-    """
-    :param url: str containing the URL to verify
-    :return: true if the URL could be a YouTube video link
-    """
-
-    return "youtu" in url
-
-
 def run(token):
-    # Suppress noise about console usage from errors
-    youtube_dl.utils.bug_reports_message = lambda: ''
-
     intents = discord.Intents.default()
     intents.message_content = True
